Package Problem
You want to send your friend a package with different things.
Each thing you put inside the package has such parameters as index
number, weight and cost.
The package has a weight limit.
Your goal is to determine which things to put into the package so that
the total weight is less than or equal to the package limit and the
total cost is as large as possible.
You would prefer to send a package which weights less in case there is
more than one package with the same price.